コンポーネント切り出し 整理
