+++ src/components/step3/Game.tsx
@@ -1,0 +1,100 @@
+
+import React , {useState} from  "react"
+import Board from  "./Board"
+
+const Game = () => {
+    const [history, setHistory] = useState([
+      {
+        squares: Array(9).fill(null),
+      },
+    ]);
+    const [stepNumber, setStepNumber] = useState(0);
+    const [xIsNext, setXIsNext] = useState(true);
+  
+    const handleClick = (i:number) => {
+      const _history = history.slice(0, stepNumber + 1);
+      const current = _history[_history.length - 1];
+      const squares = current.squares.slice();
+      if (calculateWinner(squares) || squares[i]) {
+        /*勝利条件に該当する場合 || OかXかいずれかがある場合*/
+        return;
+      }
+      squares[i] = xIsNext ? "X" : "O";
+      setHistory([
+        ...history,
+        {
+          squares: squares,
+        },
+      ]);
+      setStepNumber(_history.length);
+      setXIsNext(!xIsNext);
+    };
+  
+    const jumpTo = (step:number) => {
+      setStepNumber(step);
+      setXIsNext(step % 2 === 0);
+    };
+    const current = history[stepNumber];
+    const winner = calculateWinner(current.squares);
+    const moves = history.map((step, move) => {
+      //配列のindexが0falsyかどうかで分岐
+      console.log(move);
+      const desc = move ? "Go to move #" + move : "Go to game start";
+      return (
+        <li key={move}>
+          <button
+            onClick={() => {
+              jumpTo(move);
+            }}
+          >
+            {desc}
+          </button>
+        </li>
+      );
+    });
+  
+    let status;
+    if (winner) {
+      status = "Winner: " + winner;
+    } else {
+      status = "Next player: " + (xIsNext ? "X" : "O");
+    }
+    return (
+      <div className="game">
+        <div className="game-board">
+          <Board
+            squares={current.squares}
+            onClick={(i:number) => {
+              handleClick(i);
+            }}
+          />
+        </div>
+        <div className="game-info">
+          <div>{status}</div>
+          <ol>{moves}</ol>
+        </div>
+      </div>
+    );
+  }
+
+  function calculateWinner(squares:number[]) {
+    const lines = [
+      [0, 1, 2],
+      [3, 4, 5],
+      [6, 7, 8],
+      [0, 3, 6],
+      [1, 4, 7],
+      [2, 5, 8],
+      [0, 4, 8],
+      [2, 4, 6],
+    ];
+    for (let i = 0; i < lines.length; i++) {
+      const [a, b, c] = lines[i];
+      if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
+        return squares[a];
+      }
+    }
+    return null;
+  }
+
+  export default Game
